Move connected data to respective account views
